refactor(service): change id type to long #17

+++ src/main/java/pl/gawor/tayckner/taycknerbackend/service/service/CRUDService.java
@@ -11,7 +11,7 @@ import java.util.List;
 public interface CRUDService<Model> {
     List<Model> list();
     Model create(Model model);
-    Model read(int id);
-    Model update(int id, Model model);
-    boolean delete(int id);
+    Model read(long id);
+    Model update(long id, Model model);
+    boolean delete(long id);
 }
